HMR.
Root is no longer responsible for creating the browserHistory or the
store - this will be taken care of in the highest level index.js file.
Redux store does not replace the store by default. You also wouldn't
want the entire store to flush and reset after each hot reload.

Redbox catches errors.

Routes is now 'reactified,' that is, exported as a react element, not a
function. This way it can be used as a prop in Router (in the Root
component).

Webpack config for HMR.